Reorganize Project
* Reorganize the project in preparation for publishing to public repo

Relates to SFTPG-2630
