--- .run/[Local] PgpLambda.run.xml
@@ -1,38 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="[Local] PgpLambda" type="aws.lambda" factoryName="Local" nameIsGenerated="true">
-    <option name="credentialProviderId" value="profile:default" />
-    <option name="regionId" value="us-east-1" />
-    <option name="debugHost" value="localhost" />
-    <option name="architecture" />
-    <option name="environmentVariables">
-      <map>
-        <entry key="DECRYPTED_DONE_BUCKET" value="chris-sftpgw-pgp-done" />
-        <entry key="EncryptedSourceBucket" value="chris-sftpgw-pgp-encrypted" />
-        <entry key="PGP_KEY_LOCATION" value="chris-pgp-keys" />
-        <entry key="PGP_KEY_NAME" value="chris.asc" />
-      </map>
-    </option>
-    <option name="handler" />
-    <option name="image" value="false" />
-    <option name="logicalFunctionName" value="PgpLambda" />
-    <option name="memorySize" value="128" />
-    <option name="pathMappings">
-      <list />
-    </option>
-    <option name="platform" />
-    <option name="runtime" />
-    <option name="templateFile" value="$PROJECT_DIR$/template.yaml" />
-    <option name="timeout" value="300" />
-    <option name="useTemplate" value="true" />
-    <option name="input" value="$PROJECT_DIR$/event/s3event.json" />
-    <option name="inputIsFile" value="true" />
-    <sam>
-      <option name="additionalBuildArgs" value="" />
-      <option name="additionalLocalArgs" value="" />
-      <option name="buildInContainer" value="false" />
-      <option name="dockerNetwork" value="" />
-      <option name="skipImagePull" value="false" />
-    </sam>
-    <method v="2" />
-  </configuration>
-</component>
