+++ src/main/Azure/.dockerignore
