+++ src/main/Azure/PgpDecrypt/azure_handler.py
