+++ src/main/Azure/getting_started.md
