+++ src/main/Azure/local.settings.json
@@ -1,0 +1,13 @@
+{
+  "IsEncrypted": false,
+  "Values": {
+    "FUNCTIONS_WORKER_RUNTIME": "python",
+    "AzureWebJobsStorage": "<Connection string for Function App operational storage account>",
+    "ENCRYPTED_SOURCE_BUCKET": "<Container where GPG encrypted files will be uploaded for processing>",
+    "DECRYPTED_DONE_LOCATION": "<Container where files will be sent after processing>",
+    "AZURE_STORAGE_CONNECTION_STRING": "<Connection string for storage account that houses pre-processing, post-processing, and pgp-key containers>",
+    "PGP_KEY_LOCATION": "<Container where PGP key file is stored>",
+    "PGP_KEY_NAME": "<Name of PGP ASC key file>",
+    "PGP_PASSPHRASE": "<(optional) passphrase for pgp key file if applicable >"
+  }
+}
