+++ src/main/Azure/res/sharedconstants.py
