+++ src/main/GCP/requirements.txt
@@ -1,0 +1,10 @@
+boto3>=1.14.43
+botocore>=1.17.43
+docutils>=0.15.2
+google-cloud-storage>=2.4.0
+jmespath>=0.10.0
+python-dateutil>=2.8.1
+python-gnupg==0.4.4
+s3transfer>=0.3.3
+six>=1.15.0
+urllib3>=1.25.10
