+++ src/main/res/sharedconstants.py
